Feature: fuse adjacent memory accesses
Implements this transformation:
`SEQ(Mem(ptr + C), MEem(ptr + D)) => Mem(ptr + C)`
if C and D are adjacent memory locations.

+++ src/Arch/X86/X86Rewriter.Alu.cs
@@ -1296,7 +1296,7 @@ namespace Reko.Arch.X86
             }
         }
 
-        private Func<Expression,int,Expression> GetIncrementOperator()
+        private Func<Expression,long,Expression> GetIncrementOperator()
         {
             Constant direction = state.GetFlagGroup((uint)FlagM.DF);
             if (direction == null || !direction.IsValid)
